requirements.txt
azure-functions
azure-cosmos
azure-identity
openai
aiohttp
python-dotenv
